feat(doc): adds documentation support
* Adds organized, basic stub Markdown files for prose.
* Uses the gitbook tool for rendering prose docs locally.
* Adds a new Makefile task, `make doc` that runs the local doc server.

Addition of relevant content will come online in an ad-hoc, as needed
basis.

The GitBook tool allows custom formatting and generates static files
which can be deployed to any host in the future.

NOTE: I am not sure about GitBook, it is very annoying to customize
but it works okay for now.

SEE: https://toolchain.gitbook.com
Change-Id: Ife8b948b4e7b4631cb0c02b21bc47c3278e36612