feat(ui): Fallback Image for Alphatar
Added fallback "letter view" for Alphatars with network images in
the case the image is still being fetched from the network.

Eventually the fallback logic should be part of Flutter's image widget
itself. See: https://github.com/flutter/flutter/issues/6229

The best way to check this out is to run it on a device with
"airplane mode" turned on. The network image will not load, but the
fallback letter will still be present.

SO-14 #done

Change-Id: I7c0182938a367a5104be5e7df688cd42d3650337