feat(config): allow keeping local config file for convenience
Add more instructions on how to keep the local config file containing
the client id / secret values.

There is a template for the config file named "config.example.dart",
which gets automatically copied to "config.dart", which is actually
used by the app. Developers can locally edit the "config.dart" file to
fill in the missing values.

When the values are not defined, the gallery app now shows a simple
error screen, instead of crashing the entire app.

Change-Id: If0de9b55c4593e36c1d51ec75c72a23401f83b36