feat(auth): locally store the refresh token
Whenever a login attempt is successful, gallery now stores the refresh
token in a local storage. The stored refresh token will be used to
obtain the access token in the subsequent login attempts, so the user
does not have to perform the web-based login every single time.

TODO: the local storage will be replaced by a mock secure storage API
      written in mojom. (see SO-59).

SO-87 #done

Change-Id: I9f708228de72c584d14d86ac1f7fd69db4c7e5c6