chore(lint): exclude .mojom.dart files from static analysis
In the Makefile dart-lint target, it now passes the individual dart
file names to the analyzer, in order to exclude the .mojom.dart files
from the analysis.

The .analysis_options file gets copied to each dart package directory
so that code editors such as Atom can run the live analysis with the
desired options.

SO-97 #done

Change-Id: I5f412bbf383fb081fd7ab0c22b47533b2547c5d3