chore(config): copy the local config file when checking out the repo
The "config.dart" file used in our gallery is a file that contains the
client_id / client_secret values of the google apis project so that
gallery can talk to the actual Gmail server. This file is not checked
in, and generated by our Makefile script, so that developers can keep
this file locally and fill in the values.

This could be problematic when modules gets into the fuchsia world
build, because config.dart does not exist in the repository as is. To
get around this problem, this CL adds an update.sh to be run on jiri
update.

Change-Id: I83ff79fa4360f03c0d23befe2b96fa44d4645b47