feat(mock QB module): add a screen demonstrating the 3-column email story
Added a quick mock of the 3-column email story screen to set the
baseline of what it should look like when we're using the real modular
framework.

It's functional to some extent (e.g. selecting a thread from the list
shows the thread details on the right side), but not fully.

Screenshot:
https://fuchsia.atlassian.net/secure/attachment/11100/Screenshot_20161108-183626.png

SO-112 #done

Change-Id: I2b9526ecbad542c02d8043c71fbed6598c2a5362