feat(config): adds email JSON config
Adds tooling to enable the generation/update of a JSON configuration
file.

An OAuth token and secret are required to run `make auth`. This will
update the email/config.json file with the correct values for a self
refreshing OAuth client.

Once `make auth` is run, email/config.json can be copied to a device and
read by running applications.

In this CL:

* A make task for creating email/config.json
* A simple flutter application testing the configuration flow.
* Some format and lint issues.

Related: SO-129 Helper task for copying email/config.json to a fuchsia
device and reading configuration from dart app.

Related: SO-130 Making the email_service APIs more ergonomic.

Change-Id: Ib0d50f11322f697fe7f65308f6f87584a214cbc3