feat(email): Labels are loaded from email_session module.
The email_session state is now captured in the EmailSessionDoc which is
serialized to the bound link. Data is serialized to json and stored in
properties on a single document for now.

As a first implementatio, the email_nav module now loads its label data
from the email_session module.

Change-Id: I076ee56863cdcd53b4a3d91044fc2b2c9f893468