feat(embedded_child): generic way of adding an embedded child view
This CL introduces EmbeddedChild, which can be used to embed a child
view in a generic way. By generic, I mean that it can be used for
embedding a regular flutter widget in normal flutter application, and
embedding a ChildView of a sub-module in modular framework.

Confirmed that USPS TrackingStatus can be embedded in MessageContent
in modular and flutter.

TODO
* Make the usps module use EmbeddedChild to embed the map module.
* Add EmbeddedChildBuilder for 'youtube-video' type.

Change-Id: I0d8de056d4eaff26c98086246e4f6848bb355499