feat(widget_specs): use the element model instead of pure AST
Dart analyzer package has two separate models: abstract syntax tree
(AST), and the element model.

This CL rewrites the widget spec extractor to use the element model
instead. It's a bit more complicated to get the element model right,
but it contains a lot richer information compared to pure AST. Most
notably, the element model contains the semantic type information.

For more information, please refer to:
https://www.dartdocs.org/documentation/analyzer/0.28.1/analyzer.dart.element.element/analyzer.dart.element.element-library.html

This is a first step to determine what types of parameters are
required to construct a widget, which was not possible with the pure
AST.

Change-Id: I7276461d3a1fb4730d5630b90b8896b003d05b96