feat(widget_specs): widget size control / basic parameter list
Add width / height sliders so that users can adjust the widget size in
gallery. Also, the widget page files now provide all the parameters to
the widget constructor. Those values are still null, so the behavior
should be the same as before

SO-56 #comment widget size control / basic parameter list

Change-Id: Ibb5720ff18b42b6063220b1fc37e7adccfe6ab00