feat(widget_specs): first parameter controller: string values
As a proof of concept, added a parameter controller for string values.
For every string parameter accepted by the widget constructor, the
gallery provides an input field where the user can provide the actual
string value.

See this screenshot to get a better idea:
https://fuchsia.atlassian.net/secure/attachment/11600/Screenshot_20170123-183634.png

As it stands, this relies too much on the text-based dart code
generation, which is not desirable. Much of these features should
ideally be moved to the actual gallery code, and the generated code
should be minimized.

Will think about a better way to do this after adding a few more types
of parameter controllers.

SO-56 #comment string parameter controller

Change-Id: Idf2b5ec975054257cdb5240bd792998045a2e44d