feat(widget_specs): use @Generator annotation for model generation
This CL adds another annotation named @Generator. This can be used for
generating a model class instance that is not a primitive type. It is
meant to generate a sample model (e.g. email thread, message, user,
etc.) to be provided to a widget constructor.

@Generator annotation takes two parameters:
 - Type type:         the type of the generator, which would usually
                      be the `Fixtures` class.
 - String methodName: the name of the method to be used to generate
                      the target parameter value.

The @Generator annotation can be placed (1) in front of a model class
declaration, or (2) in front of a parameter declaration of a widget
constructor.

Example of (1):

    /// Represents a single Gmail Message
    @Generator(ModelFixtures, 'message')
    class Message {
      ...
    }

Example of (2):

    class MessageListItem extends StatelessWidget {
      final Message message;
      // ...

      MessageListItem({
        Key key,
        @required @Generator(ModelFixtures, 'message') this.message,
        ...
      })
          : super(key: key);
    }

In case the @Generator annotation appears in both places, the one on
the parameter declaration takes precedence.

There are a few restrictions of a generator class.
 - A generator class must have a default constructor.
 - The method for generating the model instance must not require any
   parameters. Having optional parameters is fine, as long as none of
   them are 'required'.

With this, the widgets gallery can now display the widgets that
require some model data (e.g. MessageListItem, etc.). The gallery page
UI also provides a button for regenerating the model.

See:
https://fuchsia.atlassian.net/secure/attachment/11700/Screenshot_20170127-214939.png

SO-176 #done

Change-Id: Ifce5b0c319f70ba32bd876edd2de17894daa70bd